+++ classes/class-simplepay-db.php
@@ -49,8 +49,9 @@ if (!class_exists('SimplePay_DB') ) {
 				simplepay_live_public_api_key VARCHAR(50),
 				simplepay_test_private_api_key VARCHAR(50),
 				simplepay_test_public_api_key VARCHAR(50),
+				simplepay_description VARCHAR(100),
 				simplepay_custom_image_url VARCHAR(100),
-				simplepay_live_mode BOOLEAN NOT NULL DEFAULT FALSE,
+				simplepay_test_mode BOOLEAN NOT NULL DEFAULT TRUE,
 				simplepay_payment_type VARCHAR(30),
 				UNIQUE KEY  id (id)
 				);";
@@ -105,9 +106,9 @@ if (!class_exists('SimplePay_DB') ) {
 			global $base_simplepay_class;
 			global $wpdb;
 
-			$cb_live_mode = 0;
-			if(isset($_POST['live_mode']) && $_POST['live_mode'] == 1) {
-			    $cb_live_mode = 1;
+			$cb_test_mode = 0;
+			if(isset($_POST['test_mode']) && $_POST['test_mode'] == 1) {
+			    $cb_test_mode = 1;
 			}    
 			
 			$wpdb->query("UPDATE " . SimplePay_DB::get_instance()->db . "
@@ -115,8 +116,9 @@ if (!class_exists('SimplePay_DB') ) {
 				simplepay_live_public_api_key = '" . $_POST['live_public_api_key'] . "',
 				simplepay_test_private_api_key = '" . $_POST['test_private_api_key'] . "',
 				simplepay_test_public_api_key = '" . $_POST['test_public_api_key'] . "',
+				simplepay_description = '" . $_POST['description'] . "',
 				simplepay_custom_image_url = '" . $_POST['custom_image_url'] . "',
-				simplepay_live_mode = '" . $cb_live_mode . "' 
+				simplepay_test_mode = '" . $cb_test_mode . "' 
 				WHERE id = '" . $base_simplepay_class->plugin_slug . "'");
 
 			wp_redirect(site_url('wp-admin/?page=simplepay-plugin'));