+++ views/admin.php
@@ -19,11 +19,10 @@ $payment_types = array('ESCROW', 'Checkout');
 	<h2>General Plugin Settings</h2>
 	<form method="post" action="admin-post.php">
 		<input type='hidden' name='action' value='simplepay_update' />
-		<div class="live-mode">
-			<h4>Live mode &nbsp;<input type="checkbox" name="live_mode" value="1" <?php echo ($form_data[0]->simplepay_live_mode == 1 ? 'checked' : '');?> /></h4>
+		<div class="test-mode">
+			<h4>Test mode &nbsp;<input type="checkbox" name="test_mode" value="1" <?php echo ($form_data[0]->simplepay_test_mode == 1 ? 'checked' : '');?> /></h4>
 		</div>
-
-			<div class="api-keys margin-top-md">
+		<div class="api-keys margin-top-md">
 			<h3>Live API Keys</h3>
 			<h4 class="no-margin-bottom">PRIVATE API Key</h4>
 			<input type="text" name="live_private_api_key" value="<?php echo $form_data[0]->simplepay_live_private_api_key; ?>" />
@@ -41,20 +40,12 @@ $payment_types = array('ESCROW', 'Checkout');
 			<input type="text" name="test_public_api_key" value="<?php echo $form_data[0]->simplepay_test_public_api_key; ?>" />
 
 			<hr class="margin-top-sm" />
-			<h3>Custom SimplePay checkout image</h3>
+			<h3>SimplePay Checkout Dialog</h3>
+			<h4>Description</h4>
+			<input type="text" name="description" value="<?php echo $form_data[0]->simplepay_description; ?>" />
+
 			<h4>Image URL</h4>
 			<input type="text" name="custom_image_url" value="<?php echo $form_data[0]->simplepay_custom_image_url; ?>" />
-
-			<!--
-			<hr class="margin-top-sm" />
-			
-			<h3>Payment Type</h3>
-			<select name="payment_type">
-				<?php // foreach ($payment_types as $type) { ?>
-						<option value="<?php // echo strtolower($type); ?>" <?php // echo ($form_data[0]->simplepay_payment_type == strtolower($type) ? 'selected' : '');?>><?php // echo $type; ?></option>
-				<?php // } ?>
-			</select>
-			-->
 		</div>
 		<div class="margin-top-md">
 			<input type="submit" name="submit" value="Save Changes" />