+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.4 (February 10, 2016)
+
+- Updated verify payment
+- Added custom SimplePay checkout description
+- Updated plugin settings page
+
 ## 1.1.3 (February 3, 2016)
 
 - Updated README